check uploading files and data with postman

+++ src/utils/cloudinary.js
@@ -2,9 +2,9 @@ import { v2 as cloudinary } from "cloudinary"
 import fs from "fs"
 
 cloudinary.config({ 
-  cloud_name: CLOUDINARY_CLOUD_NAME, 
-  api_key: CLOUDINARY_API_KEY, 
-  api_secret: CLOUDINARY_API_SECRET
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+  api_key: process.env.CLOUDINARY_API_KEY, 
+  api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
 
@@ -18,12 +18,18 @@ const uploadOnCloudinary = async (localFilePath) => {
         })
         //file has been uploaded successfull
         console.log("file is uploaded on cloudinary", res.url)
-        fs.unlinkSync(localFilePath)
-        return response;
+        if (fs.existsSync(localFilePath)) {
+            
+            fs.unlinkSync(localFilePath)
+        }
+        return res;
     }
     catch (err) {
         
-        fs.unlinkSync(localFilePath) //remove the locally saved temporary file as the upload operation got faild
+        if (fs.existsSync(localFilePath)) {
+            
+            fs.unlinkSync(localFilePath) //remove the locally saved temporary file as the upload operation got faild
+        }
         return null;
     }
 }
